channel updates
